🌱Remove api/resource/book/api.go
